simple-job-recommender-system/example/requirements.txt
streamlit==1.32.2
qdrant-client==1.8.2
sentence-transformers==2.6.1
transformers==4.38.2
torch==2.2.1
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
huggingface-hub==0.21.4
nltk==3.8.1
deep-translator==1.11.4